requirements.txt
pdf2image==1.17.0
Pillow==10.4.0
PyPDF2==3.0.1
